remove unnecessary wrapper
